+++ index.js
@@ -1,11 +1,24 @@
 import closest from 'dom-closest';
 
-export default ( selector, listener ) => ( e ) => {
-	const isDelegated = (
-		e.currentTarget.contains(closest(e.target, selector)) &&
-		e.currentTarget !== e.target
-	);
-	if ( isDelegated ) {
+const disabledElementsSelector = [
+	'input',
+	'textarea',
+	'button',
+	'select',
+	'form'
+]
+	.map((element) => `${element}[disabled]`)
+	.join(',');
+
+export default (selector, listener) => (e) => {
+	const disabledAncestor = closest(e.target, disabledElementsSelector);
+	const closestNode = closest(e.target, selector);
+	const isDelegated =
+		!e.currentTarget.contains(disabledAncestor) &&
+		e.currentTarget.contains(closestNode) &&
+		e.currentTarget !== e.target;
+	if (isDelegated) {
+		e.delegateTarget = closestNode;
 		listener(e);
 	}
 };