+++ test/index.js
@@ -3,16 +3,18 @@ import sinon from 'sinon';
 import simulant from 'simulant';
 import fn from '../index';
 
-let elements = [];
+let elements = {};
 
 before(function() {
 	window.fixture.load('/test/fixtures/index.html');
-	elements = [
-		document.querySelector('#jackie'),
-		document.querySelector('#frankie'),
-		document.querySelector('#sally'),
-		document.querySelector('#moose')
-	];
+	elements = {
+		jackie: document.querySelector('#jackie'),
+		frankie: document.querySelector('#frankie'),
+		sally: document.querySelector('#sally'),
+		moose: document.querySelector('#moose'),
+		finn: document.querySelector('#finn'),
+		baby: document.querySelector('#baby')
+	};
 });
 
 after(function() {
@@ -20,7 +22,7 @@ after(function() {
 });
 
 it('should handle delegated event', function() {
-	const [jackie, frankie, sally, moose] = elements;
+	const { jackie, frankie, sally, moose } = elements;
 
 	const callback = sinon.spy();
 	const listener = fn('#frankie', callback);
@@ -33,14 +35,17 @@ it('should handle delegated event', function() {
 	simulant.fire(sally, 'click');
 	simulant.fire(moose, 'click');
 
+	const [event] = callback.firstCall.args;
+
 	assert.equal(callback.called, true);
 	assert.equal(callback.callCount, 3);
+	assert.equal(event.delegateTarget, frankie);
 
 	jackie.removeEventListener('click', listener);
 });
 
 it('should handle unbounded delegated event', function() {
-	const [jackie, frankie, sally, moose] = elements;
+	const { jackie, frankie, sally, moose } = elements;
 
 	const callback = sinon.spy();
 	const listener = fn('#frankie', callback);
@@ -55,4 +60,28 @@ it('should handle unbounded delegated event', function() {
 
 	assert.equal(callback.called, false);
 	assert.equal(callback.callCount, 0);
+});
+
+it('shouldn’t trigger delegated event if target is inside disabled element', function() {
+	const { finn, baby } = elements;
+
+	const callback = sinon.spy();
+	const listener = fn('#baby', callback);
+
+	finn.addEventListener('click', listener);
+
+	finn.disabled = true;
+
+	simulant.fire(baby, 'click');
+	simulant.fire(baby, 'click');
+	simulant.fire(baby, 'click');
+
+	finn.disabled = false;
+
+	simulant.fire(baby, 'click');
+
+	assert.equal(callback.called, true);
+	assert.equal(callback.callCount, 1);
+
+	finn.removeEventListener('click', listener);
 });